requirements.txt
click==7.1.1
gitpython==3.1.0
python-dateutil==2.8.1
